Fix error with validation temporary table

+++ SSIS-SQL-task/SSIS-SQL-task/SqlScripts/DropTemporaryTable.sql
@@ -1,4 +1,2 @@
-DROP TABLE [TemporaryData]
-GO
-
-
+IF OBJECT_ID('dbo.TemporaryData', 'U') IS NOT NULL
+ DROP TABLE [TemporaryData];
